+++ client/cmd/dud/upload.go
@@ -365,7 +365,7 @@ func (a *app) postUpload(encryptedFile string, opts uploadOptions) ([]byte, erro
 	if err != nil {
 		return nil, err
 	}
-	defer file.Close()
+	defer func() { _ = file.Close() }()
 	info, err := file.Stat()
 	if err != nil {
 		return nil, err