+++ client/cmd/dud/v2_erase.go
@@ -336,7 +336,7 @@ func syncV2EraseDirectory(path string) error {
 	if err != nil {
 		return err
 	}
-	defer directory.Close()
+	defer func() { _ = directory.Close() }()
 	return directory.Sync()
 }